requirements.txt
click==6.6
requests==2.9.1
RPi.GPIO==0.6.2
